Remove legacy App and OutlinerMain
