+++ minders/client/legacy/components/OutlineTop.tsx
@@ -6,8 +6,8 @@ import React from 'react';
 import {StyleSheet, Text, View} from 'react-native';
 import {requireLoggedInUser} from '@toolkit/core/api/User';
 import {getChildren, hasVisibleKids, type OutlineItem} from '../model/outliner';
+import {Filters} from './AppLayout';
 import Hierarchy from './Hierarchy';
-import {Filters} from './OutlineFrame';
 import OutlineUtil from './OutlineUtil';
 import {useOutlineState, useOutliner} from './OutlinerContext';
